added multiple user capability - Backend changes

+++ backend/backend.py
@@ -1,5 +1,5 @@
 from flask import Flask, request, jsonify, stream_with_context, Response
-from flask_jwt_extended import JWTManager, create_access_token, jwt_required
+from flask_jwt_extended import JWTManager, create_access_token, jwt_required, get_jwt, get_jwt_identity, current_user
 from flask_bcrypt import Bcrypt
 from flask_cors import CORS
 import sqlite3 as sql
@@ -27,15 +27,20 @@ app.config['JWT_TOKEN_LOCATION'] = ['headers']
 jwt = JWTManager(app)
 bcrypt = Bcrypt(app)
 
+def drop_db():
+    conn = sql.connect('app.db')
+    conn.execute('DROP TABLE IF EXISTS users')
+    conn.execute('DROP TABLE IF EXISTS items')
+    conn.execute('DROP TABLE IF EXISTS chats')
+    conn.execute('DROP TABLE IF EXISTS messages')
+    conn.commit()
+
 def init_db():
     conn = sql.connect('app.db')
-    conn.execute('CREATE TABLE IF NOT EXISTS items (id INTEGER PRIMARY KEY AUTOINCREMENT, task BOOLEAN, title TEXT, content TEXT)')
-    conn.execute('CREATE TABLE IF NOT EXISTS chats (id INTEGER PRIMARY KEY AUTOINCREMENT, associatedItem INTEGER, title TEXT)')
-    conn.execute('CREATE TABLE IF NOT EXISTS messages (id INTEGER PRIMARY KEY AUTOINCREMENT, chat INTEGER, user BOOLEAN, message TEXT)')
     conn.execute('CREATE TABLE IF NOT EXISTS users (id INTEGER PRIMARY KEY AUTOINCREMENT, username TEXT, password TEXT)')
-    # hashed_password = bcrypt.generate_password_hash("password").decode('utf-8')
-    # conn.execute('INSERT INTO users (username, password) VALUES (?, ?)', ("admin", hashed_password))
-
+    conn.execute('CREATE TABLE IF NOT EXISTS items (id INTEGER PRIMARY KEY AUTOINCREMENT, user INTEGER, task BOOLEAN, title TEXT, content TEXT)')
+    conn.execute('CREATE TABLE IF NOT EXISTS chats (id INTEGER PRIMARY KEY AUTOINCREMENT, user INTEGER, associatedItem INTEGER, title TEXT)')
+    conn.execute('CREATE TABLE IF NOT EXISTS messages (id INTEGER PRIMARY KEY AUTOINCREMENT, chat INTEGER, user BOOLEAN, message TEXT)')
     conn.commit()
     
 def gpt_stream(user_prompt: str):
@@ -70,6 +75,22 @@ def chat():
         },
     )
 
+@app.route('/api/create_user', methods=["POST"])
+def create_user():
+    conn = sql.connect('app.db')
+    if request.method == "POST":    
+        try:
+            data = request.get_json()
+
+            username = data.get('username')
+            hashed_password = bcrypt.generate_password_hash(data.get('password')).decode('utf-8')
+            
+            conn.execute('INSERT INTO users (username, password) VALUES (?, ?)', (username, hashed_password))
+            conn.commit()
+            return jsonify({'message': 'User created successfully'})
+        except Exception as e:
+            return jsonify({'message': 'User creation failed: e'}), 400
+
 @app.route('/api/login', methods=["POST"])
 def login():
     conn = sql.connect('app.db')
@@ -80,7 +101,7 @@ def login():
 
     user = conn.execute('SELECT * FROM users WHERE username = ?', (username,)).fetchone()
     if user and bcrypt.check_password_hash(user[2], password):
-        access_token = create_access_token(identity=user[1])
+        access_token = create_access_token(identity=str(user[0]))
         return jsonify({'message': 'Login Success', 'access_token': access_token})
     else:
         return jsonify({'message': 'Login Failed'}), 401
@@ -88,8 +109,9 @@ def login():
 @app.route('/api/message/<id>', methods=["GET", "POST"])
 @jwt_required()
 def get_conversation(id):
+    current_user_id = get_jwt_identity()
     conn = sql.connect('app.db')
-    conversation = conn.execute('SELECT * FROM chats WHERE id = ?', (id,)).fetchone()
+    conversation = conn.execute('SELECT * FROM chats WHERE id = ? AND user = ?', (id, current_user_id)).fetchone()
 
     if request.method == "POST":
         new_message = request.get_json()["text"]
@@ -103,8 +125,9 @@ def get_conversation(id):
 @app.route('/api/response/<id>', methods=["GET", "POST"])
 @jwt_required()
 def get_response(id):
+    current_user_id = get_jwt_identity()
     conn = sql.connect('app.db')
-    conversation = conn.execute('SELECT * FROM chats WHERE id = ?', (id,)).fetchone()
+    conversation = conn.execute('SELECT * FROM chats WHERE id = ? AND user = ?', (id, current_user_id)).fetchone()
 
     if request.method == "POST":
         response = request.get_json()["text"]
@@ -118,8 +141,9 @@ def get_response(id):
 @app.route('/api/title/<id>', methods=["GET", "POST"])
 @jwt_required()
 def set_title(id):
+    current_user_id = get_jwt_identity()
     conn = sql.connect('app.db')
-    conversation = conn.execute('SELECT * FROM chats WHERE id = ?', (id,)).fetchone()
+    conversation = conn.execute('SELECT * FROM chats WHERE id = ? AND user = ?', (id, current_user_id)).fetchone()
 
     if request.method == "POST":
         title = request.get_json()["title"]
@@ -131,11 +155,12 @@ def set_title(id):
 @app.route('/api/new_chat', methods=["POST"])
 @jwt_required()
 def new_conversation():
+    current_user_id = get_jwt_identity()
     conn = sql.connect('app.db')
 
     title = "Untitled"
-    conn.execute('INSERT INTO chats (associatedItem, title) VALUES (?, ?)', (-1, title))
-    id = conn.execute('SELECT id FROM chats WHERE title = ?', (title,)).fetchall()[-1][0]
+    conn.execute('INSERT INTO chats (user, associatedItem, title) VALUES (?, ?, ?)', (current_user_id, -1, title))
+    id = conn.execute('SELECT id FROM chats WHERE title = ? AND user = ?', (title, current_user_id)).fetchall()[-1][0]
     conn.commit()
 
     messages = conn.execute('SELECT * FROM messages WHERE chat = ?', (id,)).fetchall()
@@ -145,54 +170,59 @@ def new_conversation():
 @app.route('/api/chats')
 @jwt_required()
 def get_conversations():
+    current_user_id = get_jwt_identity()
     conn = sql.connect('app.db')
-    conversations = conn.execute('SELECT * FROM chats').fetchall()
-    return jsonify([{'id': c[0], 'item': c[1], 'title': c[2]} for c in conversations])
+    conversations = conn.execute('SELECT * FROM chats WHERE user = ?', (current_user_id,)).fetchall()
+    return jsonify([{'id': c[0], 'item': c[2], 'title': c[3]} for c in conversations])
 
 @app.route('/api/chats/delete/<id>', methods=["GET", "POST"])
 @jwt_required()
 def delete_chat(id):
+    current_user_id = get_jwt_identity()
     conn = sql.connect('app.db')
     if request.method == 'POST':
-        conn.execute('DELETE FROM chats WHERE id = ?', (id,))
+        conn.execute('DELETE FROM chats WHERE id = ? AND user = ?', (id, current_user_id))
         conn.commit()
     return jsonify({'message': 'Chat deleted successfully'})
 
 @app.route('/api/items/<tasks>', methods=["GET", "POST"])
 @jwt_required()
 def notes(tasks):
+    current_user_id = get_jwt_identity()
     conn = sql.connect('app.db')
     tasks = False if tasks == "False" else True
 
     if request.method == 'POST':
         data = request.get_json()
-        conn.execute('INSERT INTO items (task, title, content) VALUES (?, ?, ?)', (tasks, data['title'], data['content']))
+        conn.execute('INSERT INTO items (user, task, title, content) VALUES (?, ?, ?, ?)', (current_user_id, tasks, data['title'], data['content']))
         conn.commit()
 
-    notes = conn.execute('SELECT * FROM items WHERE task = ?', (tasks,)).fetchall()
-    notes_list = [{'id': row[0], 'task': row[1], 'title': row[2], 'content': row[3]} for row in notes]
+    notes = conn.execute('SELECT * FROM items WHERE user = ? AND task = ?', (current_user_id, tasks)).fetchall()
+    notes_list = [{'id': row[0], 'task': row[2], 'title': row[3], 'content': row[4]} for row in notes]
     return jsonify(notes_list)
 
 @app.route('/api/items/update/<id>', methods=["GET", "POST"])
 @jwt_required()
 def update_note(id):
+    current_user_id = get_jwt_identity()
     conn = sql.connect('app.db')
     if request.method == 'POST':
         data = request.get_json()
         # First get the current task status
-        current_note = conn.execute('SELECT task FROM items WHERE id = ?', (id,)).fetchone()
+        current_note = conn.execute('SELECT task FROM items WHERE id = ? AND user = ?', (id, current_user_id)).fetchone()
         if current_note:
-            conn.execute("UPDATE items SET title = ?, content = ? WHERE id = ?", 
-                        (data['title'], data['content'], id))
+            conn.execute("UPDATE items SET title = ?, content = ? WHERE id = ? AND user = ?", 
+                        (data['title'], data['content'], id, current_user_id))
             conn.commit()
     return []
 
 @app.route('/api/items/delete/<id>', methods=["GET", "POST"])
 @jwt_required()
 def delete_note(id):
+    current_user_id = get_jwt_identity()
     conn = sql.connect('app.db')
     if request.method == 'POST':
-        conn.execute('DELETE FROM items WHERE id = ?', (id,))
+        conn.execute('DELETE FROM items WHERE id = ? AND user = ?', (id, current_user_id))
         conn.commit()
     return jsonify({'message': 'Note deleted successfully'})
